requirements.txt
requests
requests_toolbelt
jinja2
openpyxl
imapclient
pyyaml
tornado
pymysql
sqlalchemy
apscheduler
wxpy
numpy
pandas
bs4